pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "trade-agsi"
version = "0.1.0"
description = "Replication code for treaty-level AGSI construction and panel validation."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "scipy>=1.10",
  "pandas>=2.0",
  "scikit-learn>=1.3",
  "openpyxl>=3.1",
  "pyyaml>=6.0",
  "statsmodels>=0.14",
  "linearmodels>=5.3",
]

[project.optional-dependencies]
umap = ["umap-learn>=0.5"]
hdbscan = ["hdbscan>=0.8"]
pot = ["POT>=0.9"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
